LayoutTests/http/tests/xmlhttprequest/redirect-cross-origin-2-expected.txt

CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/reply.xml due to access control checks.
CONSOLE MESSAGE: Not allowed to use restricted network port 22: http://localhost:22/
Test that a cross-origin redirect to a server that responds is indistinguishable from one that does not. Should say PASS:

PASS
